+++ .github/issue_template.md
@@ -1,0 +1,24 @@
+<!-- 
+By filing an Issue, you are expected to comply with the Code of Conduct, 
+including treating everyone with respect.
+
+Not all sections will apply to all issue types.
+-->
+
+### Description
+[Detailed description of the problem and the impact it has]
+
+### Steps to Reproduce (Bug)
+[Please be as specific as possible]
+
+**Expected behavior:** [What you expect to happen]
+
+**Actual behavior:** [What actually happens]
+
+**Frequency:** [How regularly does it occur?]
+
+### Versions (Add all that apply)
+* Java version: [`java -version`]
+* OS Name & Version: [`cat /etc/*release`]
+* Docker Version: [`docker version`]
+* Cloud VM, type, size: [Amazon Web Services I3-large]
