+++ .github/pull_request_template.md
@@ -1,0 +1,7 @@
+<!-- Thanks for sending a pull request! Please check out our contribution guidelines: -->
+
+## PR description
+
+## Fixed Issue(s)
+<!-- Please link to fixed issue(s) here using format: fixes #<issue number> -->
+<!-- Example: "fixes #2" -->
