+++ .github/workflows/build.yml
@@ -1,32 +1,33 @@
 name: Build and Publish
 
 on:
-  push:
-    branches:
-      - main
   pull_request:
+  push:
     branches:
       - main
 
 jobs:
   build-and-publish:
     runs-on: ubuntu-latest
-
+    if: ${{ github.actor != 'dependabot[bot]' }}
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-
-      - name: Set up JDK 17
-        uses: actions/setup-java@v2
+      - name: Checkout Repo
+        uses: actions/checkout@v3
+      - name: Set up Java
+        uses: actions/setup-java@v3
         with:
-          java-version: "17"
-          distribution: "adopt"
-
+          distribution: adopt
+          java-version: 17
+          cache: gradle
+      - name: spotless
+        run: ./gradlew --no-daemon --parallel clean spotlessCheck
       - name: Build with Gradle
         run: ./gradlew build
-
       - name: Publish to Artifactory
-        run: ./gradlew artifactoryPublish
+        uses: gradle/gradle-build-action@v2
+        if: contains('refs/heads/release-', github.ref) || github.ref == 'refs/heads/main'
         env:
-          ARTIFACTORY_USER: ${{ secrets.ARTIFACTORY_USER }}
-          ARTIFACTORY_KEY: ${{ secrets.ARTIFACTORY_KEY }}
+          ARTIFACTORY_USER: ${{ secrets.ARTIFACTORY_USER}}
+          ARTIFACTORY_KEY: ${{ secrets.BESU_ARTIFACTORY }}
+        with:
+          arguments: --no-daemon --parallel publish artifactoryPublish --scan
