+++ .github/workflows/dco-merge-group.yml
@@ -1,0 +1,10 @@
+name: dco
+on:
+  merge_group:
+
+jobs:
+  dco:
+    runs-on: [besu-research-ubuntu-8]
+    if: ${{ github.actor != 'dependabot[bot]' }}
+    steps:
+      - run: echo "This DCO job runs on merge_queue event and doesn't check PR contents"
