+++ .github/workflows/dco.yml
@@ -1,0 +1,20 @@
+name: dco
+on:
+  pull_request:
+  workflow_dispatch:
+
+jobs:
+  dco:
+    runs-on: [besu-research-ubuntu-8]
+    if: ${{ github.actor != 'dependabot[bot]' }}
+    steps:
+      - run: echo "This DCO job runs on pull_request event and workflow_dispatch"
+      - name: Get PR Commits
+        id: 'get-pr-commits'
+        uses: tim-actions/get-pr-commits@v1.2.0
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+      - name: DCO Check
+        uses: tim-actions/dco@v1.1.0
+        with:
+          commits: ${{ steps.get-pr-commits.outputs.commits }}
