+++ .github/workflows/repolinter.yml
@@ -1,0 +1,24 @@
+# SPDX-License-Identifier: Apache-2.0
+# Hyperledger Repolinter Action
+name: Repolinter
+
+on:
+  workflow_dispatch:
+  push:
+    branches:
+      - master
+      - main
+  pull_request:
+    branches:
+      - master
+      - main
+
+jobs:
+  build:
+    runs-on: [besu-research-ubuntu-16]
+    container: ghcr.io/todogroup/repolinter:v0.10.1
+    steps:
+      - name: Checkout Code
+        uses: actions/checkout@v3
+      - name: Lint Repo
+        run: bundle exec /app/bin/repolinter.js --rulesetUrl https://raw.githubusercontent.com/hyperledger-labs/hyperledger-community-management-tools/main/repo_structure/repolint.json --format markdown
