+++ .gitignore
@@ -1,46 +1,43 @@
-.gradle
-build/
-!gradle/wrapper/gradle-wrapper.jar
-!**/src/main/**/build/
-!**/src/test/**/build/
-
-### IntelliJ IDEA ###
-.idea/modules.xml
-.idea/jarRepositories.xml
-.idea/compiler.xml
-.idea/libraries/
-*.iws
+*.bak
+*.swp
+*.tmp
+*~.nib
 *.iml
-*.ipr
-### IntelliJ IDEA ###
-out/
-!**/src/main/**/out/
-!**/src/test/**/out/
-
-### Eclipse ###
-.apt_generated
+*.launch
+*.log
+.lh/*
+db/
+db.version
 .classpath
-.factorypath
+.DS_Store
+.gradletasknamecache
+.externalToolBuilders/
+.gradle/
+.idea/
+.loadpath
+.metadata
+.prefs
 .project
+.recommenders/
 .settings
 .springBeans
-.sts4-cache
+.vertx
 bin/
-!**/src/main/**/bin/
-!**/src/test/**/bin/
-
-### NetBeans ###
-/nbproject/private/
-/nbbuild/
-/dist/
-/nbdist/
-/.nb-gradle/
-
-### VS Code ###
-.vscode/
-
-### Mac OS ###
-.DS_Store
-
-### Gradle ###
-.gradle/
+local.properties
+target/
+tmp/
+**/build/
+out/
+*.vscode/
+gradle/flow/
+*.rlib
+*.d
+*.a
+*.class
+scripts/.java-version
+**/src/jmh/generated_tests/
+**/src/jmh/generated/*
+hs_err_pid*
+/shomei.db
+node_modules
+/**/build/
