+++ gradle/versions.gradle
@@ -1,0 +1,51 @@
+/*
+ * Copyright Hyperledger Besu Contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+dependencyManagement {
+    dependencies {
+        dependency "com.google.guava:guava:31.1-jre"
+
+        dependency 'net.java.dev.jna:jna:5.13.0'
+
+        dependency 'org.hyperledger.besu:ipa-multipoint:0.8.3-SNAPSHOT'
+
+        dependency 'org.assertj:assertj-core:3.24.2'
+
+        dependencySet(group: 'io.tmio', version: '2.4.2') {
+            entry 'tuweni-bytes'
+            entry 'tuweni-rlp'
+            entry 'tuweni-units'
+        }
+
+        dependencySet(group: 'com.google.errorprone', version: '2.18.0') {
+            entry 'error_prone_annotation'
+            entry 'error_prone_check_api'
+            entry 'error_prone_core'
+            entry 'error_prone_test_helpers'
+        }
+
+        dependencySet(group: 'org.junit.jupiter', version: '5.8.2') {
+            entry 'junit-jupiter'
+            entry 'junit-jupiter-api'
+            entry 'junit-jupiter-engine'
+            entry 'junit-jupiter-params'
+        }
+
+        dependencySet(group:'com.fasterxml.jackson.core', version:'2.14.2') {
+            entry 'jackson-databind'
+        }
+
+    }
+}
