+++ src/test/java/org/hyperledger/besu/ethereum/trie/verkle/NodeUpdaterMock.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright Besu Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ */
 package org.hyperledger.besu.ethereum.trie.verkle;
 
 import java.util.HashMap;
@@ -7,17 +22,18 @@ import org.apache.tuweni.bytes.Bytes32;
 
 public class NodeUpdaterMock implements NodeUpdater {
 
-    public HashMap<Bytes, Bytes> storage;
-    
-    public NodeUpdaterMock() {
-        this.storage = new HashMap<Bytes, Bytes>();
-    }
+  public HashMap<Bytes, Bytes> storage;
 
-    public NodeUpdaterMock(HashMap<Bytes, Bytes> storage) {
-        this.storage = storage;
-    }
+  public NodeUpdaterMock() {
+    this.storage = new HashMap<Bytes, Bytes>();
+  }
 
-    public void store(Bytes location, Bytes32 hash, Bytes value) {
-        storage.put(location, value);
-    }
+  public NodeUpdaterMock(HashMap<Bytes, Bytes> storage) {
+    this.storage = storage;
+  }
+
+  @Override
+  public void store(Bytes location, Bytes32 hash, Bytes value) {
+    storage.put(location, value);
+  }
 }
